+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- (dataset/problem_definition) add plaid version to infos.
 - (sample) add `memory_safe` to save to run pyCGNS save in a subprocess and prevent memory leaks.
 - (sample) add `del_feature` method and rename `_add_feature` without leading underscore.
 - (docs) explain release process in Contributing page.